+++ lib/configs/recommended.ts
@@ -8,10 +8,12 @@ export const configRecommended = {
     'top/no-top-level-side-effects': [
       'error',
       {
+        allowDerived: false,
         allowedCalls: ['Symbol'],
         allowedNews: ['Map', 'Set', 'WeakMap', 'WeakSet'],
+        allowFunctionProperties: false,
         allowIIFE: false,
-        allowDerived: false
+        allowPropertyAccess: true
       }
     ],
     'top/no-top-level-variables': [